requirements.txt
uvicorn==0.24.0  # Понижаем версию для совместимости
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==4.6.0  # Возвращаемся к версии 4.6.0 для совместимости с fastapi-cache2
celery==5.4.0
flower==2.0.1
fastapi-cache2[redis]==0.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
apscheduler==3.10.4
asyncpg==0.29.0
alembic==1.14.1
gunicorn==21.2.0
pydantic==2.10.6
starlette==0.45.3
bcrypt==4.0.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.2
httpx==0.28.1
locust==2.31.8
setuptools

# uvicorn==0.34.0
# sqlalchemy==2.0.37
# psycopg2-binary==2.9.10
# python-dotenv==1.0.1
# redis==4.6.0
# celery==5.4.0
# flower==2.0.1
# fastapi-cache2[redis]==0.2.1
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.20
# email-validator==2.2.0
# apscheduler==3.10.4
# asyncpg==0.29.0
# alembic==1.14.1
# gunicorn==21.2.0
# pydantic==2.10.6
# starlette==0.45.3
# bcrypt==4.0.1